requirements.txt
fastapi
uvicorn
sqlalchemy
openai
nltk
pydantic
python-dotenv
httpx
pytest
pytest-asyncio
